% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{analyze_ggpairs}
\alias{analyze_ggpairs}
\title{Analyze the results from a ggpairs,}
\usage{
analyze_ggpairs(data, alpha = 0.05, message = TRUE)
}
\arguments{
\item{data}{the df that will be using}

\item{alpha}{the level of significance}

\item{message}{if you want to return a result}
}
\description{
Analyze the results from a ggpairs,
}
\note{
prints the regressors that best fit a normal distribution. prints the regressors that are corrolated with the dependent variable. prints most corrolated variable between regressors
}
