% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_anova.R
\name{anova_test_manual}
\alias{anova_test_manual}
\title{manual creation of Anova test.}
\usage{
anova_test_manual(SST, SSTR, nT, k, alpha = 0.05, message = FALSE)
}
\arguments{
\item{SST}{the total sum of squares}

\item{nT}{the total number of observations for every treatment}

\item{k}{the number of levels or factors in the treatment}

\item{alpha}{the rejection level}

\item{message}{False if you want the conclusion hypothesis, TRUE if you you don't want the conclusion hypothesis.}

\item{SSR}{the total sum of error squares}
}
\value{
The Anova table or the conclusion for the hypothesis.
}
\description{
Manually creates the table for the anova table using the defined parameters.
}
