% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{BreushPaga}
\alias{BreushPaga}
\title{Test if the elpsilons have variance equals to sigma^2 using BreushPaga}
\usage{
BreushPaga(updatedModel, alpha = 0.05)
}
\arguments{
\item{updatedModel}{the model to test the significance}

\item{alpha}{the level of significance}
}
\description{
Test if the elpsilons have variance equals to sigma^2 using BreushPaga
}
