% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{check_model_significance}
\alias{check_model_significance}
\title{To test if the model for the f test is significant.}
\usage{
check_model_significance(model, alpha)
}
\arguments{
\item{model}{the model to test the significance}

\item{alpha}{the level of significance}
}
\description{
To test if the model for the f test is significant.
}
