% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{compareRsquare}
\alias{compareRsquare}
\title{Compare the Rsquare with Adjusted R-square}
\usage{
compareRsquare(model, threshold = 0.1)
}
\arguments{
\item{data}{the data frame for the model to implement}

\item{threshhold}{the required distance for between the Rsquare and adjRsquare}
}
\description{
Compare the Rsquare with Adjusted R-square
}
