% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{compute_sums_squares}
\alias{compute_sums_squares}
\title{computes the sum squares for a model}
\usage{
compute_sums_squares(model)
}
\description{
computes the sum squares for a model
}
