% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{create_interaction_terms}
\alias{create_interaction_terms}
\title{prints all the interactions terms from a data set, excluding the first column which is treated as the dependent variable.}
\usage{
create_interaction_terms(data)
}
\arguments{
\item{data}{the data to use}
}
\description{
prints all the interactions terms from a data set, excluding the first column which is treated as the dependent variable.
}
