% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{csvPath}
\alias{csvPath}
\title{Read a CSV File}
\usage{
csvPath(path, removeSpace = TRUE)
}
\arguments{
\item{path}{A character string specifying the path to the CSV file.}

\item{removeSpace}{A logical indicating whether to remove spaces in column names (default is TRUE).}
}
\value{
A dataframe containing the contents of the CSV file.
}
\description{
This function reads a CSV file and loads it as a dataframe.
}
