% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Compute Descriptive Statistics}
\usage{
descriptive_stats(data)
}
\arguments{
\item{data}{A dataframe containing the data.}
}
\value{
A dataframe containing descriptive statistics for numeric variables.
}
\description{
Computes basic descriptive statistics (count, mean, median, SD, min, max, IQR) for numeric columns.
}
