% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{descriptive_stats_by_factor}
\alias{descriptive_stats_by_factor}
\title{Compute Descriptive Statistics by Factor}
\usage{
descriptive_stats_by_factor(data, factor_col)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{factor_col}{A character string specifying the factor column to group by.}
}
\value{
A dataframe containing descriptive statistics for numeric variables grouped by the factor.
}
\description{
Computes descriptive statistics for numeric variables grouped by a factor column.
}
