% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{DurbinWatson}
\alias{DurbinWatson}
\title{Test if the epsilons are independent using DurbinWatson}
\usage{
DurbinWatson(updatedModel, alpha = 0.05)
}
\arguments{
\item{updatedModel}{the model to test the significance}

\item{alpha}{the level of significance}
}
\description{
Test if the epsilons are independent using DurbinWatson
}
