% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{Eepsilon}
\alias{Eepsilon}
\title{Test if the epsilons have expectation 0}
\usage{
Eepsilon(updatedModel, categorical = TRUE, alpha = 0.05, message = TRUE)
}
\arguments{
\item{updatedModel}{the model to test the significance}

\item{categorical}{if any of the variables is categorical}

\item{alpha}{the level of significance}

\item{message}{if you want to return a string that prints the conclusion}
}
\description{
Test if the epsilons have expectation 0
}
