% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.barplot.R
\name{effect}
\alias{effect}
\title{A function that illustrates bar plot of tw o variables}
\usage{
effect(data, x, y, color)
}
\arguments{
\item{data}{the data selected}

\item{x}{numeric. numeric vector}

\item{y}{numeric. numeric vector}

\item{color}{color of the bar plot}
}
\value{
bar plot a bar plot is returned with selected color
}
\description{
The function below will show us the bar plot of two variables and how they depend on each other
}
\author{
Islam Islamov and Leyla Ellazova
}
