% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{excelPath}
\alias{excelPath}
\title{Read an Excel File}
\usage{
excelPath(path, removeSpace = TRUE, sheet = NULL)
}
\arguments{
\item{path}{A character string specifying the path to the Excel file.}

\item{removeSpace}{A logical indicating whether to remove spaces in column names (default is TRUE).}

\item{sheet}{An optional sheet name or index to read (default is NULL, which reads the first sheet).}
}
\value{
A dataframe containing the contents of the Excel sheet.
}
\description{
This function reads an Excel file and loads it as a dataframe.
}
