% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_anova.R
\name{flsd_CI}
\alias{flsd_CI}
\title{Fisher Least Square Difference, direct confidence interval for two specific levels.}
\usage{
flsd_CI(modelAnova, levelVar, group1, group2, alpha = 0.05, message = FALSE)
}
\arguments{
\item{group2}{first group}

\item{alpha}{the confidence level for rejection.}

\item{message}{TRUE if you want to see a message, FALSE if you just want the CI}

\item{name}{description}
}
\description{
Fisher Least Square Difference, direct confidence interval for two specific levels.
}
