% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{fullModelCreation}
\alias{fullModelCreation}
\title{Create a Full Linear Model with All Numeric Predictors}
\usage{
fullModelCreation(data, y)
}
\arguments{
\item{data}{A dataframe containing the dataset.}

\item{y}{A character string specifying the dependent variable.}
}
\value{
A fitted linear model (\code{lm} object).
}
\description{
Builds a linear model using all numeric predictors in a dataset.
}
