% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{fullModelSummary}
\alias{fullModelSummary}
\title{returns the model with the summary}
\usage{
fullModelSummary(data, y)
}
\arguments{
\item{data}{the data frame for the model to implement}

\item{y}{the dependent variable}
}
\description{
returns the model with the summary
}
