% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graphs.R
\name{histogram_normal_residuals}
\alias{histogram_normal_residuals}
\title{Historgram for the normal fitted residuals}
\usage{
histogram_normal_residuals(model, binwidth = 0.5, bins = 5)
}
\arguments{
\item{model}{model to use}
}
\description{
creates a histrogram fitted with a normal for the residuals.
}
