% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{listExcelSheets}
\alias{listExcelSheets}
\title{List Sheets in an Excel File}
\usage{
listExcelSheets(path)
}
\arguments{
\item{path}{A character string specifying the path to the Excel file.}
}
\value{
A character vector containing the names of the sheets in the Excel file.
}
\description{
This function lists all the sheet names in an Excel file.
}
