% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{modelObject}
\alias{modelObject}
\title{returns the model as an object}
\usage{
modelObject(data, y, x = list(), interaction = list(), summary = FALSE)
}
\arguments{
\item{data}{the data frame for the model to implement}

\item{y}{the dependent variable}

\item{x}{the list of all the regressors}

\item{interaction}{which regressors will have interaction}

\item{summary}{if you wish to include the summary of the model}
}
\description{
returns the model as an object
}
