% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{My package}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{x must be a matrix, containing the observations of interest}

\item{mean}{mean is the vector of the multivariate gaussian distribution}

\item{varcovM}{variance covariance matrix (squared matrix)}

\item{Log}{by default, Log = T}
}
\value{
A list with the matrix and the value of the density of a multivariate normal distribution
}
\description{
description
}
\details{
détails
}
\examples{
x <- 1
}
