% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{shapiroTest}
\alias{shapiroTest}
\title{shapiroTest to test normality}
\usage{
shapiroTest(updatedModel, alpha = 0.05)
}
\arguments{
\item{updatedModel}{the model to test the significance}

\item{alpha}{the level of significance}
}
\description{
shapiroTest to test normality
}
