% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_anova.R
\name{singleAOV}
\alias{singleAOV}
\title{Anova with just one factor.}
\usage{
singleAOV(df, treatment_col, value_col, summary = FALSE)
}
\arguments{
\item{df}{data frame}

\item{treatment_col}{the treatment coloumn's name (where the treatments name are)}

\item{value_col}{the observations coloumn from the levels.}
}
\description{
Anova with just one factor.
}
