% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_anova.R
\name{SSE}
\alias{SSE}
\title{Compute the sum of squares due to error (SSE)}
\usage{
SSE(df, treatment_col, value_col, message = FALSE)
}
\arguments{
\item{df}{the data frame.}

\item{treatment_col}{the column with the treatment names}

\item{value_col}{the column with the values.}

\item{message}{TRUE if you want to see the result, FALSE if the you don't want the result.}
}
\description{
Compute the sum of squares due to error (SSE)
}
