% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_linearRegression.R
\name{compute_sums_squares}
\alias{compute_sums_squares}
\title{Compute Sum of Squares for a Linear Model}
\usage{
compute_sums_squares(model)
}
\arguments{
\item{model}{A linear model object (\code{lm}).}
}
\value{
A list with SST, SSR, and SSE.
}
\description{
Computes SST (Total Sum of Squares), SSR (Regression Sum of Squares), and SSE (Error Sum of Squares).
}
